Add support for column partitioning

+++ src/main/scala/glint/messages/server/request/PullMatrix.scala
@@ -6,4 +6,4 @@ package glint.messages.server.request
   * @param rows The row indices
   * @param cols The column indices
   */
-private[glint] case class PullMatrix(rows: Array[Long], cols: Array[Int]) extends Request
+private[glint] case class PullMatrix(rows: Array[Long], cols: Array[Long]) extends Request
