+++ src/main/scala/glint/models/client/retry/RetryBigMatrix.scala
@@ -16,7 +16,7 @@ import scala.concurrent.{Future, ExecutionContext}
 class RetryBigMatrix[@specialized V](underlying: BigMatrix[V], val attempts: Int = 3) extends BigMatrix[V] {
 
   val rows: Long = underlying.rows
-  val cols: Int = underlying.cols
+  val cols: Long = underlying.cols
 
   /**
     * Pulls a set of rows
@@ -55,7 +55,7 @@ class RetryBigMatrix[@specialized V](underlying: BigMatrix[V], val attempts: Int
     * @return A future containing either the success or failure of the operation
     */
   override def push(rows: Array[Long],
-                    cols: Array[Int],
+                    cols: Array[Long],
                     values: Array[V])(implicit ec: ExecutionContext): Future[Boolean] = {
     implicit val success = Success[Boolean](identity)
     retry.Directly(attempts) { () =>
@@ -72,7 +72,7 @@ class RetryBigMatrix[@specialized V](underlying: BigMatrix[V], val attempts: Int
     * @return A future containing the values of the elements at given rows, columns
     */
   override def pull(rows: Array[Long],
-                    cols: Array[Int])(implicit ec: ExecutionContext): Future[Array[V]] = {
+                    cols: Array[Long])(implicit ec: ExecutionContext): Future[Array[V]] = {
     implicit val success = Success.always
     retry.Directly(attempts) { () =>
       underlying.pull(rows, cols)
