+++ src/test/scala/glint/matrix/GranularBigMatrixSpec.scala
@@ -15,7 +15,7 @@ class GranularBigMatrixSpec extends FlatSpec with SystemTest with Matchers {
         val model = client.matrix[Double](1000, 1000)
         val granularModel = new GranularBigMatrix[Double](model, 10000)
         val rows = new Array[Long](1000000)
-        val cols = new Array[Int](1000000)
+        val cols = new Array[Long](1000000)
         val values = new Array[Double](1000000)
         var i = 0
         while (i < rows.length) {
@@ -43,7 +43,7 @@ class GranularBigMatrixSpec extends FlatSpec with SystemTest with Matchers {
         val model = client.matrix[Double](1000, 1000)
         val granularModel = new GranularBigMatrix[Double](model, 10000)
         val rows = new Array[Long](1000000)
-        val cols = new Array[Int](1000000)
+        val cols = new Array[Long](1000000)
         val values = new Array[Double](1000000)
         var i = 0
         while (i < rows.length) {
